Stop mypy from re-checking numpy's own bundled stubs
The CI matrix's Python 3.10 job resolves numpy 2.2.6 (the newest
numpy release that still ships a cp310 wheel), whose inline stubs
trip hundreds of mypy 2.3.0's stricter checks (stale type: ignore
comments, overload-overlap) under our project-wide strict = true.
follow_imports = "silent" didn't suppress these consistently across
numpy's submodules; ignore_errors = true on the numpy override does,
without touching the type information mypy infers from those stubs.

Also fixes a genuine type mismatch this surfaced: eigvalsh's return
dtype is only pinned to float64 on newer numpy, so pin it explicitly
to match the codebase's existing convention (selector.py, expansion.py).

+++ src/sexpansion/expansion.py
@@ -189,7 +189,7 @@ class ExpandedAlgebra:
 
     def eigenvalues(self) -> FloatArray:
         """Eigenvalues of the (symmetric) Cartan-Killing metric, ascending."""
-        result: FloatArray = np.linalg.eigvalsh(self.cartan_killing_metric())
+        result: FloatArray = np.linalg.eigvalsh(self.cartan_killing_metric()).astype(np.float64)
         return result
 
     def eigenvectors(self) -> tuple[FloatArray, FloatArray]:
